Preorder Traversal
We first take an explicit stack data structure and push the root node to it.(if the root node is not NULL).
Then we use a while loop to iterate over the stack till the stack remains non-empty.
In every iteration we first pop the stack’s top and print it. 
Then we first push the right child of this popped node and then push the left child, if they are not NULL. We do so because stack is a last-in-first-out(LIFO) data structure. We need to access the left child first, so we need to push it at the last.
The execution continues and will stop when the stack becomes empty. In this process, we will get the preorder traversal of the tree.