Create Print Pascal’s Triangle
First, we will run a loop(say row) from 1 to n.
We will use a loop(say col) to iterate over each column i.e. from 1 to n. And inside the nested loops, we will do the following steps:
First, we will consider row-1 as n and col-1 as r.
After that, we will simply calculate the value of the combination using a loop. 
The loop will run from 0 to r. And in each iteration, we will multiply (n-i) with the result and divide the result by (i+1).
Finally, we will add the element to a temporary list.
After calculating all the elements for all columns of a row, we will add the temporary list to our final answer list.
Finally, we will return the answer list.